Remove Books and Links pages.
Books are moved to highlights.yashagarwal.in.
I did not really utilize links page. It feels like a burden that I have
to save the links. However, it is important that I come up with another
system to store and revisit the links that I read on a daily basis.